requirements.txt
langchain
langchain_community
langchain_groq
langchain-openai
langchain-google-genai
uvicorn
python-dotenv
pandas
numpy
pypdf
faiss-cpu
structlog
PyMuPDF
-e .
streamlit
pytest
langchain-core[tracing]